feat: optimize data types for settlement amounts and loop counters (#12)

+++ src/interfaces/ITributeDraft.sol
@@ -5,23 +5,26 @@ pragma solidity ^0.8.27;
 /// @notice Aggregates multiple Consumption Units into a single Tribute Draft
 interface ITributeDraft {
     struct TributeDraftEntity {
+        /// @notice Unique ID of the tribute draft
         bytes32 tributeDraftId;
-        // owner of all aggregated consumption units
+        /// @notice Owner of the tribute draft
         address owner;
-        // currency shared across all consumption units (ISO 4217)
+        /// @notice Numeric currency code using ISO 4217
         uint16 settlementCurrency;
-        /// ISO 8601 format: 20250923
+        /// @notice Worldwide day in compact format (e.g., 20250923)
         uint32 worldwideDay;
-        // aggregated settlement amount
-        uint256 settlementAmountBase;
-        uint256 settlementAmountAtto;
-        // source CU ids
+        /// @notice Aggregated amount expressed in natural units (base currency units).
+        uint64 settlementAmountBase;
+        /// @notice Aggregated amount expressed in fractional units (atto, 1e-18). Must satisfy 0 <= amount < 1e18.
+        uint128 settlementAmountAtto;
+        /// @notice Hashes identifying linked consumption units
         bytes32[] cuHashes;
+        /// @notice Timestamp when the tribute draft was submitted
         uint256 submittedAt;
     }
 
     event Submitted(
-        bytes32 indexed tdId, address indexed owner, address indexed submittedBy, uint256 cuCount, uint256 timestamp
+        bytes32 indexed tdId, address indexed owner, address indexed submittedBy, uint32 cuCount, uint256 timestamp
     );
 
     error EmptyArray();
